selesaikan kode tugas 6

+++ main/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseNotFound, HttpResponseRedirect
 from main.forms import ProductForm
 from django.urls import reverse
 from django.http import HttpResponse
@@ -12,6 +12,7 @@ import datetime
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import get_object_or_404, redirect
+from django.views.decorators.csrf import csrf_exempt
 
 # Create your views here.
 
@@ -33,6 +34,10 @@ def show_json_by_id(request, id: int):
     data = Product.objects.filter(pk=id)
     return HttpResponse(serializers.serialize('json', data), content_type='application/json')
 
+def get_product_json(request):
+    product_item = Product.objects.all()
+    return HttpResponse(serializers.serialize('json', product_item))
+
 @login_required(login_url='/login')
 def show_main(request):
     products = Product.objects.filter(user=request.user)
@@ -45,6 +50,22 @@ def show_main(request):
 }
 
     return render(request, "main.html", context)
+
+@csrf_exempt
+def add_product_ajax(request):
+    if request.method == 'POST':
+        name = request.POST.get("name")
+        price = request.POST.get("price")
+        description = request.POST.get("description")
+        user = request.user
+
+        new_product = Product(name=name, price=price, description=description, user=user)
+        new_product.save()
+
+        return HttpResponse(b"CREATED", status=201)
+
+    return HttpResponseNotFound()
+
 
 def create_product(request):
     form = ProductForm(request.POST or None)
@@ -108,28 +129,95 @@ def add_amount(request, product_id):
 
     return render(request, "main.html", context)
 
-
-def add_amount(request, id):
-    current_product = Product.objects.get(pk=id)
-    current_product.amount += 1
-    current_product.save()
-    return HttpResponseRedirect(reverse('main:show_main'))
-
+@csrf_exempt
 def decrement_amount(request, id):
     current_product = Product.objects.get(pk=id)
     current_product.amount -= 1
     if current_product.amount == 0 :
         return delete_product(request, id)
     current_product.save()
-    return HttpResponseRedirect(reverse('main:show_main'))
+    return HttpResponse(b"DECREMENTED PRODUCT SUCCESSFULLY", status=200)
 
 
+@csrf_exempt
 def delete_product(request, id):
-    current_product = Product.objects.get(pk=id)
-    current_product.delete()
+    if request.method == 'DELETE':
+        current_product = Product.objects.get(pk=id)
+        current_product.delete()
+        return HttpResponse(b"DELETED PRODUCT SUCCESSFULLY", status=200)
     return HttpResponseRedirect(reverse('main:show_main'))
 
+@csrf_exempt
+def add_amount(request, id):
+    current_product = Product.objects.get(pk=id)
+    current_product.amount += 1
+    current_product.save()
+    return HttpResponse(b"ADDED PRODUCT SUCCESSFULLY", status=200)
 
+@csrf_exempt
+def edit_product(request, id):
+    # Get product berdasarkan ID
+    product = Product.objects.get(pk = id)
+
+    # Set product sebagai instance dari form
+    form = ProductForm(request.POST or None, instance=product)
+
+    if form.is_valid() and request.method == "POST":
+        # Simpan form dan kembali ke halaman awal
+        form.save()
+        return HttpResponseRedirect(reverse('main:show_main'))
+
+    context = {'form': form}
+    return render(request, "edit_product.html", context)
+
+
+def get_product_json(request):
+    product_item = Product.objects.filter(user=request.user)
+    return HttpResponse(serializers.serialize('json', product_item))
+
+@csrf_exempt
+def add_product_ajax(request):
+    if request.method == 'POST':
+        name = request.POST.get("name")
+        price = request.POST.get("price")
+        amount = request.POST.get("amount")
+        description = request.POST.get("description")
+        user = request.user
+
+        new_product = Product(name=name, price=price, amount=amount, description=description, user=user)
+        new_product.save()
+
+        return HttpResponse(b"CREATED", status=201)
+
+    return HttpResponseNotFound()
+
+@csrf_exempt
+def edit_product_ajax(request, id):
+    if request.method == 'POST':
+        name = request.POST.get("name")
+        price = request.POST.get("price")
+        amount = request.POST.get("amount")
+        description = request.POST.get("description")
+        user = request.user
+
+        new_product = Product(name=name, price=price, amount=amount, description=description, user=user)
+        product = Product.objects.get(pk=id)
+        product = new_product
+        product.save()
+
+        return HttpResponse(b"CREATED", status=201)
+
+    return HttpResponseNotFound()
+
+@login_required(login_url='/login')  
+def show_json_by_user(request):
+    data = Product.objects.filter(user=request.user)
+    return HttpResponse(serializers.serialize("json", data), content_type="application/json")
+
+@login_required(login_url='/login')  
+def show_xml_by_user(request):
+    data = Product.objects.filter(user=request.user)
+    return HttpResponse(serializers.serialize("xml", data), content_type="application/xml")
